[26] refactor: change project structure (#57)
* refactor: change project structure

* refactor: up styles folder to src level
