+++ src/App.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/no-unresolved */
 import React, { ReactElement } from 'react';
 import { Route, Routes } from 'react-router-dom';
-import RequireAuth from 'hocs/RequireAuth';
+import RequireAuth from 'core/hocs/RequireAuth';
 import BoardsPage from 'pages/BoardsPage/BoardsPage';
 import HomePage from 'pages/HomePage/HomePage';
 import NotfoundPage from 'pages/NotFoundPage/NotFoundPage';
